cleand up code and added page fault handler
