+++ kernel/src/memory/memory.rs
@@ -1,8 +1,13 @@
 use x86_64::{
-    structures::paging::{FrameAllocator, Mapper, Page, OffsetPageTable, PageTable, PhysFrame, Size4KiB},
+    structures::paging::{FrameAllocator, Mapper, OffsetPageTable, Page, PageTable, Size4KiB},
     VirtAddr,
-    PhysAddr,
 };
+
+pub unsafe fn init(hhdm_base: VirtAddr) -> OffsetPageTable<'static> {
+    let pml4 = active_level_4_table(hhdm_base);
+
+    OffsetPageTable::new(pml4, hhdm_base)
+}
 
 unsafe fn active_level_4_table(physical_memory_offset: VirtAddr) -> &'static mut PageTable {
     use x86_64::registers::control::Cr3;
@@ -16,14 +21,8 @@ unsafe fn active_level_4_table(physical_memory_offset: VirtAddr) -> &'static mut
     &mut *page_table_ptr
 }
 
-pub unsafe fn init(hhdm_base: VirtAddr) -> OffsetPageTable<'static> {
-    let pml4 = active_level_4_table(hhdm_base);
-
-    OffsetPageTable::new(pml4, hhdm_base)
-}
-
 /// Creates an example mapping for the given page to frame `0xb8000`.
-pub fn create_example_mapping(
+pub fn create_mapping(
     page: Page,
     mapper: &mut OffsetPageTable,
     frame_allocator: &mut impl FrameAllocator<Size4KiB>,
